+++ grafana-plugin/src/models/schedule/schedule.types.ts
@@ -120,6 +120,7 @@ export interface Layer {
 export interface ShiftEvents {
   shiftId: string;
   events: Event[];
+  priority: number;
   isPreview?: boolean;
 }